Move provisions model to joda-money

+++ src/main/resources/sql/busfest-schema.sql
@@ -8,7 +8,7 @@ DROP TABLE IF EXISTS conventions;
 CREATE TABLE provisions (
   id BIGINT(20) NOT NULL AUTO_INCREMENT,
   description VARCHAR(255) NOT NULL,
-  cost INT NOT NULL,
+  cost DECIMAL(13,2) NOT NULL,
   convention_id BIGINT(20) DEFAULT NULL,
   created_at DATETIME NOT NULL,
   updated_at DATETIME NOT NULL,
